+++ app/components/science/index.ts
@@ -5,6 +5,6 @@ export { default as ScienceNootropics } from "./ScienceNootropics";
 export { default as ScienceEducation } from "./ScienceEducation";
 export { default as ScienceExplainer } from "./ScienceExplainer";
 export { default as RealisedSolution } from "./RealisedSolution";
-export { default as SynergyChart } from "./SynergyChart";
 export { default as ScienceDifferent } from "./ScienceDifferent";
-export { default as EvidenceLadder } from "./EvidenceLadder";
+export { default as ScienceProof } from "./ScienceProof";
+export { default as ScienceCTA } from "./ScienceCTA";